update: reflect dom changes to dataDomain
