+++ src/VGesture.ts
@@ -1,23 +1,20 @@
-import Fastdom from 'fastdom';
-import fastdomPromiseExtension from 'fastdom/extensions/fastdom-promised'
+
 import { HandDetector } from './models/HandDetector';
 import { Camera } from './models/Camera';
 import { DataDomain } from './models/DataDomain';
 import { GestureManager } from './models/GestureManager';
 import { AbstractGesturePlugin } from './plugins/Plugin';
 import { VGestureError } from './error';
-import { traverse } from './utils/dom/traverse';
 import { error, warn } from './utils/console'
 import { CANVAS_ELEMENT_ID, LEFT_HAND_CONTAINER_ELEMENT_ID, RIGHT_HAND_CONTAINER_ELEMENT_ID, VIDEO_ELEMENT_ID, WRAPPER_ELEMENT_ID } from './constant'
 import { ERROR_TYPE, SESSION_STATE, Handedness } from './types'
 import type { OperationKey } from './Gestures/Gesture';
-import type { Boundary2D, ElementBoundary, OperationRecord, Helper, VGestureOption } from './types';
+import type { OperationRecord, Helper, VGestureOption } from './types';
 import { Stage } from './models/Stage';
 import { register } from './utils/prebuilt';
 import { isValidPlugin } from './utils/validation/plugin';
 
 
-const fastdom = Fastdom.extend(fastdomPromiseExtension);
 const $$driverKey = Symbol('driverKey');
 
 
@@ -29,12 +26,14 @@ export class VGesture {
   private detector: HandDetector | null = null;
   private camera: Camera | null = null;
   private stage: Stage | null = null;
+  private observer: MutationObserver;
 
   private sessionState: SESSION_STATE;
   private frameId: number | null = null;
   //VGestureConfig
   public dataDimension!: 2; // currently only 2 is allowed.
   public helper: Helper | null;
+
 
   constructor(options?: VGestureOption) {
     // populating configs
@@ -49,6 +48,14 @@ export class VGesture {
     this.dataDimension = dataDimension;
     this.gestureManager = new GestureManager()
     this.sessionState = SESSION_STATE.IDLE;
+    this.observer = new MutationObserver(() => {
+      this.flush()
+    })
+    this.observer.observe(document.body, {
+      childList: true,
+      subtree: true,
+      characterData: true
+    })
   }
 
   async initialize() {
@@ -71,10 +78,27 @@ export class VGesture {
 
     await this.detector.initialize();
 
+    // this.observer.observe(document.body, {
+    //   childList: true,
+    //   subtree: true,
+    //   characterData: true,
+    // })
     this.initialized = true;
     this.sessionState = SESSION_STATE.READY;
   }
 
+  /**
+   * Since mutation observer works only for DOM changes, its difficult to catch whethere
+   * there was reflow at vgesturable elements via cssom changes. 
+   * 
+   * So its highly recommended to not to change vgesturable element's position after initialize.
+   * 
+   * But in case need of recalculating elements position, use this function to refresh positions.
+   * e.g) language change for global website, responsive website etc..
+   */
+  async flush() {
+    this.gestureTargetCollection.update();
+  }
   async startDetection() {
     if (!this.initialized || !this.detector) {
       throw new VGestureError(ERROR_TYPE.VALIDATION, 'VGesture.startDetection', 'Validation Error: V-Gesture not initialized')
@@ -104,6 +128,7 @@ export class VGesture {
     this.stage!.disconnect()
     this.gestureManager.disposeAll();
     this._cleanStartedElems();
+    this.observer.disconnect();
     this.initialized = false;
   }
 
@@ -217,42 +242,9 @@ export class VGesture {
   }
 
   private async _generateGestureTargetCollection() {
-    const PREFIX = 'vgesturable'
-    const elemBoundaries: ElementBoundary[] = []
-    let id = 0;
 
-    await fastdom.mutate(() => {
-      // traverse from  Dom tree, rooting from body node, find all elems with gClickable specified elements
-      // create kdtree to handle event target domain
-      traverse(document.body, (elem) => {
-        if ((elem as HTMLElement).hasAttribute('vgesturable')) {
-          const clickableElem = elem as HTMLElement
-          const { top, left, width, height } = clickableElem.getBoundingClientRect();
-          let elemId = clickableElem.id;
-
-          if (!elemId) {
-            elemId = `${PREFIX}-${id}`
-            id++;
-          }
-
-          clickableElem.id = elemId;
-
-          const x = left + width / 2;
-          const y = top + height / 2;
-          const dx = width / 2;
-          const dy = height / 2;
-          const boundary = [x, y, dx, dy] as Boundary2D;
-          const ElementBoundary = {
-            id: elemId,
-            dimension: boundary.length / 2,
-            boundary
-          }
-
-          elemBoundaries.push(ElementBoundary)
-        }
-      })
-      this.gestureTargetCollection = new DataDomain(elemBoundaries);
-    })
+    this.gestureTargetCollection = new DataDomain([]);
+    this.gestureTargetCollection.update()
   }
 
   private _createStarterElems() {
